+++ packages/akashic-cli-export/src/html/convertBundle.ts
@@ -57,7 +57,7 @@ export async function promiseConvertBundle(options: ConvertTemplateParameterObje
 		try {
 			options.sandboxConfigJsCode = readSandboxConfigJs(options.source);
 			sandboxConfig = require(path.join(options.source, "sandbox.config.js"));
-		} catch (error) {
+		} catch (_error) {
 			throw Error("failed read sandbox.config.js.");
 		}
 		validateSandboxConfigJs(sandboxConfig, options.autoSendEventName, options.autoGivenArgsName);