+++ packages/akashic-cli-scan/eslint.config.cjs
@@ -1,0 +1,19 @@
+const eslintConfig = require("@akashic/eslint-config");
+
+module.exports = [
+    ...eslintConfig,
+    {
+        files: ["src/**/*.ts"],
+        languageOptions: {
+            sourceType: "module",
+            parserOptions: {
+                project: "tsconfig.json"
+            },
+        },
+        ignores: [
+          "**/*.js",
+          "*.cjs",
+          "**/__tests__/**/*"
+        ]
+    }
+];