+++ packages/akashic-cli-serve/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 2.0.4
+
+### Patch Changes
+
+- [#1564](https://github.com/akashic-games/akashic-cli/pull/1564) [`495f94f`](https://github.com/akashic-games/akashic-cli/commit/495f94fd53cf9c7f60028591c78e0b33b4f41893) Thanks [@ShinobuTakahashi](https://github.com/ShinobuTakahashi)! - Fix bug in replay mode with -A option in serve
+
+## 2.0.3
+
+### Patch Changes
+
+- [#1554](https://github.com/akashic-games/akashic-cli/pull/1554) [`68a2f92`](https://github.com/akashic-games/akashic-cli/commit/68a2f929a9fc04967833b610ca82d201df60b122) Thanks [@yu-ogi](https://github.com/yu-ogi)! - feat(serve): add support for playlog duration
+
+- [#1553](https://github.com/akashic-games/akashic-cli/pull/1553) [`1165411`](https://github.com/akashic-games/akashic-cli/commit/1165411463a9567bc853a2e0c93b4a69f6f1ae62) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@akashic/headless-driver` to `2.17.4`.
+
 ## 2.0.2
 
 ### Patch Changes